Refactor URL parsing logic into a utility function and streamline catalog path handling in Tree component
