+++ src/pages/MapViewer/components/Toolbox/components/ToolboxItems/index.tsx
@@ -16,6 +16,9 @@ const ToolboxItems = () => {
     actions: { setActivePage, returnResultsPage },
   } = useToolbox();
 
+  // console.log('Selected Collection', selectedCollection);
+  // console.log('Selected Collection Items', selectedCollectionItems);
+
   const {
     state: { activeFilters },
     actions: { setResultsPage },
